Tweak Heka plugin_chansize and others
Speculatively tweak our Heka configuration to see if it helps to prevent
the `nginx_access_stat` StatFilter plugin from becoming stuck. By
'stuck' I mean that the filter's buffer (or channel, in Heka-speak) is
full and it is unable to process further log entries.

Note this is not intended as a permanent change, rather a way to help
identify the underlying cause.

These changes are entirely speculative. They aim to resolve an issue
where we are not receiving Nginx logs or metrics from the following
machines to our central Heka and Graphite services on `logging-1` and
`graphite-1` respectively:

> backend-1_backend_production
> backend-2_backend_production
> backend-3_backend_production
> backend-lb-1_backend_production
> backend-lb-2_backend_production
> cache-1_router_production
> cache-2_router_production
> cache-3_router_production
> frontend-lb-1_frontend_production
> frontend-lb-2_frontend_production

The problem occurs in Staging and in Production but not in Preview,
presumably due to the low traffic in the Preview environment.

I'm increasing `plugin_chansize` based on a comment from one of the Heka
developers in this thread:

https://mail.mozilla.org/pipermail/heka/2014-August/000171.html

> There's a static pool of packs, and if those packs are exhausted then
> you're going to deadlock b/c everybody will be waiting for new packs and
> none are being freed.  Similarly, if the traffic flowing through the
> router backs up, then things will come to a stop b/c, well, just about
> everything flows through the router.
>
> [...]
>
> So if you really have a case where one plugin is going to need to
> consume a large number of packs at a single time, then you can crank the
> number of available packs way up. This will increase the resident memory
> size of the running hekad process, but it will give you the headroom you
> need to not cause everything else to block due to running out of packs.

I'm reducing the ticket interval on the `StatAccumInput` plugin to tick
every second on the basis that it might help to process data more
quickly.