Put nginx_access_stats ticker interval back to 5s
Partially revert c4a374bd0, which set the `ticker_interval` for the
`nginx_access_stats` Heka plugin to 1 second.

The `ticker_interval` determines how often metrics are flushed to Carbon
Cache, so this interval should be equal to or greather than the
interval for Carbon's highest resolution data (5 seconds):
https://github.gds/gds/puppet/blob/b15c13527283292a4355264a42c34f5204ea4ce9/modules/govuk/files/node/s_graphite/storage-schemas.conf#L8-L10

Otherwise, our metric counts are lower than they should be as Graphite
does not process all of the metrics we send it.

There's some good information about this here:
https://kevinmccarthy.org/blog/2013/07/18/10-things-i-learned-deploying-graphite/#make-sure-that-your-statsd-flush-interval-is-at-least-as-long-as-your-graphite-interval