Mongo backup: Don't log to a file or to email
This script had a lot of code to allow you to have the Mongo backup logs
sent to an email address or a file.

Some of that code is quite poorly written (e.g. nested if statements
evalutating the same condition) and the I/O redirection is difficult to
follow.

Since we don't need to log to an email address, rip this out in a bid to
simplify the script.