Check mongodump exit code and exit early if fails
Explicitly check the exit code returned by `mongodump`, and exit the
backup script early if `mongodump` returns a non-zero exit code.

Explicitly check the exit code because the `errexit` option (`set -e`)
doesn't apply to commands called inside functions.

Exit the script rather than return `1` from the function because there's
not much point in continuing if one of the backups has failed. Also,
even if I replaced `mongodump` with `false` in my testing and `dbdump()`
returns `1`, the rest of the script continued as if nothing had failed.
So the return code didn't seem to affect the outcome of the script.