Remove parser_type setting for LogstreamerInput
The Heka `LogstreamerInput` plugin no longer takes a parser type as of
verion 0.9.0.

The setting was removed in this commit:
https://github.com/mozilla-services/heka/commit/1b211c67108c390df907d9ae667f1c2467446238

Looking at the source code for version 0.8.3, the default `parser_type`
for LogstreamerInput was `token`, meaning that this should not affect
version 0.8.3 in Production:
https://github.com/mozilla-services/heka/blob/v0.8.3/plugins/logstreamer/logstreamer_input.go#L88

I tested this change in Vagrant with version 0.8.3 and version 0.9.0 and
it does not cause any errors.