Increase heka's `pool_size` to 130
The default pool size for heka is 100. This means we can have 100
packs (messages) for all of the inputs to use. But we might be using all
of those packs through the `StatFilter` so none of them will be left for
the `StatAccumInput` to use.

Increasing this will mean that even if the *match* chan and the *in*
chan for the `StatFilter` and the `TcpOutput` are completely full, there
will still be some packs left for the `StatAccumInput` to use.