Remove legacy alias for bouncer.${app_domain}
We should be okay to trust DNS to resolve bouncer.${app_domain} across all
environments now that we have external IP addresses presented to the vShield
Edge in all environments.

Removing the legacy alias here, and allowing DNS to do its thing, fixes an
alert for Bouncer in Preview. This should also be okay to roll out to every
environment without detrimental effect.