Restrict elasticsearch transport to cluster members
At the moment the module opens the elasticsearch transport port (9300)
to all. This should only ever be used by other cluster members, so this
updates it to add rules for the specific cluster members only.

Note to do this, I've added a mechanism that looks up the IP addresses
for given hostnames by looking for their corresponding govuk::host
instance.