Bump puppet-graphite dependency to 2.0.3
To fix the install problem described in:

- https://github.com/gds-operations/puppet-graphite/commit/84011b8d34ebc0028bc230437531054960a5da87
- https://www.pivotaltracker.com/story/show/85067926

I've removed the exact version number from `Puppetfile` and am letting
`Puppetfile.lock` pin the version, because the number isn't that important
to us, what is important is that it's currently at the latest and that it
doesn't change unless we want it to.

I haven't changed the version of puppet-python, which librarian attempted to
update automatically, because I'm not able to extensively test that at the
moment. The new version of puppet-graphite doesn't depend on a later version
of puppet-python though.