Link to opsmanual docs for app health check alerts
So that it's easy to find the documentation for dealing with Whitehall and
Publisher scheduled publishing/editions alerts. For 2nd line Pivotal
[#91594558] and [#87306190] and builds upon gds/opsmanual#461.

Because this resource is created dynamically for all apps that request it
we'll have to be careful about matching the sub-section titles in the
opsmanual. I've split the variables out, in part, to make that clear.

I started writing tests for this and then remembered that we can't test
exported resources with rspec-puppet :(