Heka: Temporarily change CarbonOutput matcher
Following advice from Rob Miller at Mozilla, we added a new sandboxed
Lua plugin for processing Nginx request metrics in the previous commit.

That plugin sets the payload type to `txt` and payload name to
`statmetric` [1], which means that our existing Carbon output [2] does
not match those payloads since it expects the type to be `statmetric`.

While we clarify with Rob Miller how we can improve this, add a new
`CarbonOutput` section to our Heka config which will match the payloads
generated by the Lua plugin. This will temporarily replace our existing
configuration for the Carbon output [2], which we are not using for
anything else.

[1]: https://gist.github.com/rafrombrc/3858a7885e766fbefa61#file-http_req_stat_filter-lua-L99
[2]: https://github.com/alphagov/puppet-heka/blob/8eb97c0ce1d02ef103c54b7e2c8538b58e2b6ac1/templates/etc/heka/graphite.toml.erb