Upgrade NodeJS dependency in the dev VM
The 'govuk_frontend_toolkit' test suite requires 'npm'.
'npm' is installed with recent versions of 'nodejs', but the the Precise
version of the dev VM installs a very old version of 'nodejs' without
'npm'.

The 'nodejs' dependency is also set in 'common.yaml' and 'common.precise.yaml'.
We didn't want to change those versions because that would upgrade the production
versions of node, potentially breaking the things that use it.