Add enabled parameter to mongodb::backup
For some node/environment combinations we want to disable
MongoDB backups. For example, exception-handler-1 in staging
generates a lot of exceptions during the production data sync.
These exceptions are mostly noise, and we don't care about backing
them up.

This commit is a no-op by default, but adds a parameter that we
can set in hieradata to disable the backup.