Fix statsd metric for Puppet runs
When I added this metric in 2c23250, I mistakenly understood the `s`
parameter to represent 'seconds'.

Statsd timers only accept milliseconds as a unit [1]. Statsd was in fact
interpreting the `s` as an indication that I was logging a 'set' metric
[2], leading to metrics being created as:

> stats.sets.api-1_api_production_.puppet.run_duration.count

I'm not sure why there's a trailing underscore in the hostname, but I'll
tackle that separately if it's still an issue once statsd starts
interpreting these metrics as timers.

[1]: https://github.com/etsy/statsd/blob/23b331895cc4b22b64a19fd0e7b6def6f6f30d9e/docs/metric_types.md#timing
[2]: https://github.com/etsy/statsd/blob/23b331895cc4b22b64a19fd0e7b6def6f6f30d9e/docs/metric_types.md#sets