Add path to Airbrake endpoint
Neither the Logrus Airbrake hook [1] nor the underlying Airbrake library
[2] it uses add the path when sending a notice to an Airbrake-compatible
API.

I think this should be fixed upstream in the Airbrake library [2], but
for now we can work around this by explicitly setting the path.

Without this, exceptions are logged to `/` and so Errbit redirects the
Airbrake client to Signon.

[1]: https://github.com/Sirupsen/logrus/blob/55eb11d21d2a31a3cc93838241d04800f52e823d/hooks/airbrake/airbrake.go#L29
[2]: https://github.com/tobi/airbrake-go/blob/5b5e269e1bc398d43f67e43dafff3414a59cd5a2/airbrake.go#L19