Fix trailing underscore in Puppet metrics
This metric, which records the Puppet run duration, had a trailing
underscore after the hostname:

  stats.timers.cache-1_router_production_.puppet.run_duration

This was caused by the use of `hostname -A`, which returns all fully
qualified domain names for the machine it runs on. Unfortunately,
presumably because this option can return multiple values, its output
includes a trailing space even if just one FQDN is returned.

Use `hostname --fdqn` instead, which will return just one value with no
trailing space (aside from a newline, which will be truncated from the
subshell output). This option is deprecated according to the manpage for
`hostname, but I'm not too worried since the man page was last updated
in 2009 and the option still exists.

I'm not worried about moving the old metrics as we only have about 4
days of data for this metric and the old metrics can be cleaned up later
using this Fabric task:

https://github.com/alphagov/fabric-scripts/blob/master/graphite.py#L5-L8

[1]: http://manpages.ubuntu.com/manpages/trusty/man1/hostname.1.html