Automatically removed unused kernels
Replace `govuk_unused_kernels` with the now open-source gem that does the
same thing and has tests. I've pinned the version because it has the
potential to do harm and we'd want to test new releases.

Now that we have greater confidence that it's doing the right thing we can
run it automatically from cron rather than running it reactively using
Fabric whenever we get disk space alerts.

As such it no longer makes sense to link the alert about low space on
`/boot` to the opsmanual entry about running `govuk_unused_kernels`.

When merged I'll need to update:

 - [ ] gds/opsmanual
 - [ ] alphagov/fabric-scripts

Short explanation of the `cron.daily` script:

- `set -e -o pipefail` ensure that we bail out if `ubuntu_unused_kernels`
  returns a non-zero exit code. We shouldn't trust passing STDOUT to
  `apt-get` if this happens.
- `PATH` is set because we can't be sure what cron will provide us. This
  list of paths covers everything that we need.
- `xargs -r` ensures that `apt-get` isn't run if the list of package is
  empty. This accounts for the newline character that
  `ubuntu_unused_kernels` outputs regardless.
- `apt-get -y` accepts any prompts. It wouldn't matter here, but as a
  sidenote, it's not possible to have `apt-get` prompt when you're feeding
  it STDIN from `xargs`.
- It will only produce STDOUT when packages are being removed or APT
  produces an error. This will go to the `machine.email` list, which I think
  is desirable (at least for now).