Fix govuk::app service removal.
This was using the pattern param to the service type, which meant that
it would kill any process that matched the app's title. This was overly
loose and has resulted in this killing unrelated processes (at the
moment, the 'search' app is being removed, and this is also killing
elasticsearch on the dev VM).