Upgrade Node.js to 0.10.37
StatsD is currently the only thing using Node.js in our stack.
Tests for StatsD run against 0.10, and in the version of StatsD
newer than the one we are running, support for 0.6 was dropped.

Performance Platform's Node.js apps are going to start running on
the GOV.UK stack, so it makes sense to upgrade the version of
Node we're using.

Eventually we might want something to manage multiple Node versions
on a single machine, but I don't think that's something we need now.