Create spotlight_proxy vhost
Create a new vhost for Spotlight - spotlight_proxy - in order to enable a
smoother deployment of Spotlight on to GOV.UK infrastructure given that the
application already has a pre-existing vhost on frontend-lb which
proxies through to the soon-to-be-defunct Performance Platform
infrastructure.

By doing this, we're removing the pain of having to work around and tidy up
a lot of things when Spotlight is deployed to the GOV.UK frontend machines.

Once this has been added, we can update the Router to use the new
spotlight_proxy vhost to route traffic to Performance Platform
infrastructure. When Spotlight is deployed to GOV.UK, a new vhost for
spotlight will be created to route traffic to GOV.UK infrastructure. After
this point, we can then remove spotlight_proxy.