Add performance platform frontend apps to lb
This is feature flagged due to the existing proxy to performance
platform infrastructure, which would clash with the new vhost that this
will create. Once the router has been upgraded to point to the
spotlight-proxy backend, we can safely enable the spotlight app. It is
also feature flagged so we don't add the vhost on environments that do
not have the new performance platform boxes