Add performance platform bigscreen view
This is a purely static app used for displaying big screen views of
performance platform dashboards. Following the approach taken for
performance platform notifier, just the parts of govuk::app that are
required for a static vhost are used.

The port number we specify is not actually used - however the current
nginx vhost define requires it is specified, and the logic requiring an
app_proxy to be present is very intertwined with a standard vhost.

The reason static hosting works like this is because the proxy vhost
first tries looking for files in the document root, if these do not
exist it will send to the upstream proxy