Extract out draft content store setup into a class
https://trello.com/c/2ZsPnGKo

setting up `Govuk::App::Envvar` defaults was becoming confusing
when this logic was in `govuk::apps::content_store`. a separate
class ensures that we're setting up the correct defaults for that
class's scope.

this change enables us to remove a parameter, and selectively
enable draft mode in development by simply including that class
only in s_development.pp instead.