Require index name to delete Elasticsearch index
By default, Elasticsearch allows indices to be deleted without explicilty
specifying a name. This allows for API calls such as `curl -XDELETE
'http://localhost:9200/*'` or `curl -XDELETE 'http://localhost:9200/_all'`
to respond positively and delete all indexes.

Obviously, this is a bit unsafe. It's not something we ever want to do.
Intent should always be explicit, so if you really do want to delete an
index, you should have to specify the index name a la `curl -XDELETE
'http://localhost:9200/mainstream'`.

Thankfully, we can enforce the use of a name in this way using the
`action.disable_delete_all_indices` option in `elasticsearch.yml`, which we
pass in using the `elasticsearch/elasticsearch` Puppet module. However,
since some clusters we run are on `0.90.12` and others are on `1.4.4`, we
need to make sure we apply the right configuration given this option changed
to `action.destructuve_requires_name` in Elasticsearch versions greater
than 1.0 per https://github.com/elastic/elasticsearch/issues/4549.

As a result, this commit adds in `action.disable_delete_all_indices` and
sets it to true by default as long as the running version is greater than
the version we compare against in-line using `versioncmp`. `versioncmp` returns
-1 if the running version is less than the set baseline, a 0 if they match, and
a 1 if the running version is greater than the baseline version. As this behaviour
shouldn't ever be changed without good reason, it seems a very good idea to enforce
it in the configuration rather than allow it to be set on a per-cluster basis using a
Hiera lookup.