Add priorities for remaining directory backups
The three backups without priorities are being run in ascii-betical
order. Since the graphite backup comes first in that order and takes a
long time to run it leaves the other two backups not-run for long enough
that alerts are raised for them.

This adds explicit priorities to all three, with the graphite backup as
the lowest priority.