Upgrade logs-elasticsearch to 1.4.4
We've been running a really old version of Elasticsearch for our ELK logging
stack for quite some time, so we should probably upgrade it. We run 1.4.4 on
the API Elasticsearch cluster which is responsible for responding to search
queries from Rummager and other applications, so it's probably good to catch
up and go straight up to there.

Testing in Vagrant by importing 8500 documents from a Wikipedia dump seems
to show that, aside from babysitting the change out, everything just works
as expected. Once the VMs were reprovisioned, they started to rebalance
which took a few minutes but it can be worked around.