Switch from Oracle Java 7 to OpenJDK7
Switch from using Oracle Java 7 to OpenJDK7 in order to match what the API
Elasticsearch clusters run.

Include the govuk_java::openjdk7::jdk class because this is not included as
part of the ::jre class, as it is with the similar Oracle class.

Remove govuk_java::oracle7::jdk by ensuring => absent as we don't need this
running any more since OpenJDK7 takes its place.

Stop the setting of a default JVM using set_defaults given that we only
install one Java runtime: OpenJDK7.

This appears to work a-okay:

$ java -version
> java version "1.7.0_79"
> OpenJDK Runtime Environment (IcedTea 2.5.5) (7u79-2.5.5-0ubuntu0.12.04.1)
> OpenJDK 64-Bit Server VM (build 24.79-b02, mixed mode)