Rename PostgreSQL `master.pp` to `primary.pp`
We'd like to move away from using master/slave and instead use
primary/standby as in the PostgreSQL docs:

> Servers that can modify data are called read/write, master or
> primary servers. Servers that track changes in the master are
> called standby or slave servers.

http://www.postgresql.org/docs/9.3/static/high-availability.html

This is because the words master and slave have historically been
associated with the oppression of non-white people by white people.

This follows the lead of projects like Django which have made similar
changes:

- https://code.djangoproject.com/ticket/22667
- https://github.com/django/django/pull/2692
- https://github.com/django/django/pull/2694

This commit is the first step to doing that.