Try to be more intelligent about box_dist
If a vApp in govuk-provisioning has a template name which includes
the string 'precise' or 'trusty', we should not require people to
include it in their `nodes.local.yaml` file.

We could make this more robust based on the existing pattern for
template names, but we didn't want to make too many assumptions
about how those are structured.