Upgrade memcached module to fix deprecation warning.
Was giving:

Warning: Scope(Class[Memcached]): type() DEPRECATED: This function will
cease to function on Puppet 4; please use type3x() before upgrading to
puppet 4 for backwards-compatibility, or migrate to the new parser's
typing system.

The type function was deprecated in stdlib[1] because it's going to
become a language keyword soon.  The updated version of the memcached
module resolves this[2].

[1] https://tickets.puppetlabs.com/browse/MODULES-1473
[2] https://github.com/saz/puppet-memcached/pull/49