Disable unattended_reboot everywhere.
We've noticed some issues with kernel 3.13.0-51 on some machines.  This
resulted in some kernel Oops in the bind syscall, and caused mysql to
fail on mysql-backup-1 as well as causing issues for other processes
(fail2ban-server is a common one).

This therefore disables the automatic reboots to prevent more machines
upgrading to the new kernel until we've worked out what the issue is.