Switch to using directory environments
We have some red items in every puppet run relating to using directory
environments referring to http://links.puppetlabs.com/env-settings-deprecations

The basic details of a directory environment are that the following config vars
move to being set in `environmentpath/${environment}/environment.conf` and
optionally the manifest and module directories can be sub-dirs of the environment
directory.

We don't want that at the moment, so what I have done is set up the environment
directories, but set the manifests and modules to come from the parent directory
so that all environments share the same modules and manifests.

This has been tested with `vagrant provision` and by running `govuk_puppet` on:

1. A dev VM
2. A puppetmaster in vagrant-govuk
3. A jumpbox-2 on vagrant-govuk (required manual cert signing, but tests that
   Ubuntu 14.04 doesn't have any strange issues.