Hack the hiera config for puppet-apply
In the previous commit to enable directory environments, I removed the
`manifestdir` setting for Puppet which necessitated splitting the hiera config
into two portions, because using variable interpolation in the hiera config to
set the `:datadir:` options was no longer possible.

This commit removes that second config file, in favour of running sed inside of
`tools/puppet-apply` to create a temporary config file. This will ensure that we
don't need to maintain two config files and yet can still use `puppet-apply` on
development and testing VMs.

Thanks to @mattbostock for the suggestion!