Revert "Disable unattended_reboot everywhere."
This reverts commit 4fe5a0cd40281ff86515bc582f356bb7e032799f.

I've tested the following machines in preview:

- `jumpbox-2.management` running Trusty and package version `3.13.0-52.86`
- `mysql-backup-1.backend` running Precise and package version
  `3.13.0-52.86~precise1`

Both rebooted OK and there isn't anything suspicious in the output of dmesg.
So it should now be safe to re-enable this.

I had to run `sudo unattended-upgrade` to install the latest package on the
Trusty machine. The Precise machine had already installed it overnight. On
real machines cron will run `unattended-upgrade` before `unattended-reboot`.