Move to the latest es_dump_restore Gem
This should fix a bug with elasticsearch dump files not containing the
schema for the index being dumped.  It also allows us to update the
restore scripts to set up elasticsearch aliases correctly.

I don't believe there are any breaking changes between 0.0.8 and 1.0.0,
so this should be safe to deploy.