Add FIXME to Spotlight conditional
The spotlight_proxy class was added to move the original spotlight
class to, so that we could make the name `spotlight` available
for the application which is being migrated.

Once the application is up and running we can remove the
spotlight_proxy class and all of the logic around including it.