Change default elasticsearch shard count to 3
Previously, we'd been using the default elasticsearch shard count of 5.
However, this doesn't match well with our clusters (which are currently
all 3 node clusters); since 5 is coprime with 3, it's not possible for
shards to be evenly distributed.  We're also pretty stable in terms of
index size (ie, unlikely to need to add nodes dynamically), and rebuild
the main govuk index daily (as far as elasticsearch is concerned).

Therefore, there is no benefit in having more shards than nodes, so we
should have 3 shards.

Note that this will need careful deployment to avoid restarting multiple
nodes in elasticsearch clusters concurrently.