Use newer es_dump_restore gem for better syncing
The 1.0.0 release of es_dump_restore supports a restore_alias command,
which restores a dump to a new index, and then swaps an alias over to
point to it, much like our rummager index migration process.

This release also fixes some bugs which were causing the schema and
mappings not to be copied.

Upgrading to this release means that the indexes after syncing are
identical to those on production.

This is much improved on the previous version which would often merge
the new index with the old, or silently fail to update anything.
However, it does cause one problem: the index settings from production
are copied over, but in development we want our indexes to have 1 shard
and no replicas.  For this reason, a 1-node dev "cluster" will go into
the yellow state after running the sync now.  This can be fixed by
running rummager's `rummager:migrate_index` rake command on all indexes,
but need not be fixed in most cases, since the index will work perfectly
well in this state, for non-production purposes.

This could be fixed in future by updating `es_dump_restore` to allow
overriding some of the index settings.  This PR is still a big
improvement without this being fixed, though.