Reduce error_page_ttl from 300s to 30s
`error_page_ttl` determines how long we cache 404 pages for (I'll rename
it to be more descriptive in a separate commit).

As found during the GOV.UK Gameday 2015 in which several of our
databases were intentionally deleted in a test environment[1], and during a
recent outage in which our internal database connections failed
intermittently[2], we serve 404 pages when a database error occurs.

[1]: https://www.pivotaltracker.com/story/show/87935614
[2]: https://docs.google.com/a/digital.cabinet-office.gov.uk/document/d/1xXzW6eIfiNfuU2LXkq0c0dXjCKRseZrGq-trlrO95_8/edit?usp=sharing

Because a 404 page is a legitimate HTTP response and not an error, we do
not perform failover for 404 responses at the CDN edge.

In a bid to reduce the user impact when we suffer a database outage,
reduce the TTL for 404 pages from 5 minutes to 30 seconds.

I think we should continue to cache 404 errors because not doing so
would likely cause significant additional load on our origin; reducing
the TTL to 30 seconds seems like a reasonable compromise.