Add the RACK_MULTIPART_PART_LIMIT env var for Whitehall
A new limit was introduced in Rack 1.6 (part of the Rails 4.2 upgrade) for the number of multipart fields in a form POST. There was a bug in the code that iintroduced this in that it was including non-file fields in the limit calculation. Therefore forms with large numbers of (usually dynamically-added) fields were also being caught by the limit.

If a POST exceeds the limit, Rack simply returns a basic 500 error, with no feedback bout what is wrong.

The bug was originally thought to be fixed in Rack 1.6.1, but is now slated for release in 1.7. In the meantime, we need to disable the limit to work around the issue.

Stack Overflow post on the issue: http://stackoverflow.com/questions/27773368/rails-4-2-internal-server-error-with-maximum-file-multiparts-in-content-reached
Rack bug: https://github.com/rack/rack/pull/814
Story: https://www.pivotaltracker.com/n/projects/1261204/stories/94933134