Add env var for Errbit environment prefix
Add a new environment variable to `govuk::deploy::config` to set a
prefix that can be used by applications to know which environment to
send exceptions to.

We plan to use this to add a 'draft-' prefix to the environment used
when sending Errbit exceptions, e.g.:

https://github.gds/gds/alphagov-deployment/blob/61fcdaee3f8e788f6d2161f272a48382657c2be8/asset-manager/initializers_by_organisation/preview/airbrake.rb#L5

This is so that exceptions from the Draft stack can be sent to Errbit in
the main stack and be appropriately tagged so that we can differentiate
between the two environments.

We can filter by environment in Errbit, e.g.:
https://errbit.preview.alphagov.co.uk/apps/53cd34710da115dfc2001ca7?environment=preview