Explicitly specify content-register DB in the VM
Turns out the VM manages its database differently to the other environments. If
we want hstore to be enabled on the VM, we need to configure the database
explicitly here. We cannot rely on the app itself to enable this extension with
a database migration because the DB user is not a super-admin.