Point Backdrop and Stagecraft at specific location
Backdrop and Stagecraft work together, so need to talk to each other. The
quickest way of doing this is to point external DNS at a different location,
but since doing this would substantially break Performance Platform's
preview environment, that's not really an option.

Instead, we can achieve similar - although not exact - results using an
/etc/hosts entry or two, to point:

- stagecraft.${pp_app_domain}
- backdrop.${pp_app_domain}

to Fastly

In testing this (in development), you may find that the connection to
Fastly works fine yet you receive a 500 Domain Not Found error. This is
expected, because we aren't going to configure our Fastly service to
accept connections on the development value of ${pp_app_domain} at this
time.

This is a temporary change for testing until we migrate and update the
current performance platform dns