Fix the names used for elasticsearch indexes
Elasticsearch indexes are expected by rummager to have names of the form
`<index name>-<timestamp>-<UUID>`.

When syncing indexes, we generate such a name, but use zeros for the
UUID - mainly because there's no real need for the UUID and it's easier.
Unfortunately, I used the wrong number of zeros when setting this up.

A consequence of this is that replication hasn't been cleaning up the old
indexes (though they're left in the "closed" state, so aren't using up
RAM), so if you've run data sync a few times your elasticsearch database
directory will be getting cluttered up with old indexes.