Wrap some applications in an enabled flag
There are currently a lot of critical alerts in the draft stack for
applications which we don't expect to be running there.

This commit adds a new parameter to several of the applications that
we don't expect to run in the draft stack so that we can disable
them using environment-specific hieradata.