[Static] Allow setting of the header text
Users need a visual cue that they are looking at the Draft GOV.UK rather than
live. We'll do this by changing the header text from "GOV.UK" to "GOV.UK
(Draft)" or similar.

Successfully tested in development.

There will be a companion change in the deployment repo to set the value for
the draft stack.