Set maintenance_work_mem to 500MB for Postgresql sync
Transition's database takes a long time to sync (~50 mins), and a large part of
that time (~35 mins) is taken up by indexing large tables. (Confusingly, the
indexing starts when the progress indicator is already at 100%, making it look
as if the import has got stuck.)

Increasing maintenance_work_mem for the session from the default value (16MB)
to 500MB speeds up the indexing, making the sync a bit quicker: loading the
dump took 40m34s for me with this change, compared to 50m03s without it.

The dev VM has 4GB RAM by default, so 500MB should be fine temporarily while
replicating data. I also tried setting it to 1GB, but that only reduced the
time by a further 30s.