Enable conditional logstreaming for Signon
This is the first of a series of commits to re-enable logstreaming for
Signon.

Make logstreaming configurable so we can enable it in preview only, to
see whether we're leaking anything. Once we have confidence we're not
leaking anything, we can revert this change and remove the `logstream`
argument to the `govuk::app` class call.