Increase etcd heartbeat interval
Our network latency on Skyscape, being a virtualised environment, is
very variable.

This would appear to cause heartbeat timeouts, i.e.:

    [etcd] May 29 09:48:05.070 INFO      | etcd-1.management.production: warning: heartbeat time out peer="etcd-2.management.production" missed=1 backoff="2s"

I hope that by tweaking the heartbeat interval and election timeout we
can prevent these errors from occurring.

The default for heartbeat interval is 50ms, and the default for the
election timeout is 200ms. The etcd documentation recommends setting the
latter to 4 or 5 times the value of the former:

https://coreos.com/docs/cluster-management/debugging/etcd-tuning/#time-parameters