Replace users and groups with hieradata usernames
This commit replaces the concept of users and groups in our Puppet code
with just a list of users that can be pulled in from hieradata.

I'm not sure that the concept of 'groups' in our Puppet code really adds
anything, and could potentially be confused with Unix groups.

There's some duplication in the staging and production hieradata with
`ssh::config::allow_users` which could potentially be removed in the
future.