Give Backdrop Write its own RabbitMQ virtual host
The Transformer Worker uses Celery as its RabbitMQ client.

Celery is very opinionated about declaring its own exchanges and queues.
It tries to declare the following exchanges and queues:

- celeryev
- celery
- reply.celery.pidbox
- ef741ca7-c1bb-3bb3-9571-98171b945fa8.reply.celery.pidbox

...and so on.

At first, we tried creating the `celeryev` exchange for Celery, but it
tries to re-declare the exchange anyway.

We looked at the `auto_declare` option, but found little documentation
on it. Ultimately, we decided to create a separate virtual host for
Backdrop Write/Transformer Worker so that various Celery instances would
not conflict with each other.

We no longer declare the `backdrop_write` exchange ourselves as this is
created for us. Based on the configuration here:
https://github.com/alphagov/pp-puppet/blob/c5d5cae22339dc805827de23ce033db9a68c49a6/modules/performanceplatform/manifests/pp_rabbitmq.pp