Support app can serve generated CSVs
Adds an internal location in nginx for /csvs, which maps to
`/data/uploads/support-api/csvs`.

The support API will dump CSVs in that path over NFS, and the support
app will ask nginx to serve them via `send_file`, which sends an
`X-Accel-Redirect` header.

Ensure that the CSV target directory exists and is writeable by the
support-api app, so make `deploy` own it.