Use open source module for unattended reboots
Use our open source Puppet module for unattended reboots [1], which
moves all of the logic not specific to GOV.UK to a reusable module.

The `govuk_unattended_reboot` class remains but now acts as a wrapper
for the open source module and applies configuration specific to GOV.UK.

Advantages to using the new module include:

- Errors are now logged to rsyslog using the `logger` command, meaning
  we're no longer logging to a file that needs to be tailed by
  logship (nor use a logstream resource).

- Errors that occur in the `post-reboot-unlock` Upstart task are now
  logged to rsyslog for easier troubleshooting.

- The new module supports executing scripts before a node is rebooted,
  to take a node out of a cluster or load balancer pool, for example.

Functionally, the reboot mechanism is the same.

One small difference in the open source module is that it does not
include the change in 6e3faefe to release the reboot lock as part of the
cronjob; instead it relies on the Upstart task to do this. Should we see
the same issues again we should be able to used the improved logging
output to better diagnose the issue and finder a cleaner solution.

[1]: https://github.com/alphagov/puppet-unattended_reboot