Provide default for 'type' param; update docs
When using govuk_rabbitmq::exchange, it is necessary to provide a value to
the type parameter even if you are using 'absent' as the value of the
'ensure' parameter. You can therefore end up in daft situations where you
are trying to remove an exchange, yet have to provide a value to the type
param.

To stop this occurring, provide a default undef value to the type.

Also adds some class documentation.