replication: ignore transition db by default.
Given transition isn't being actively developed, and its database is
large, this updates the default ignore list to skip the transition
database by default.  This also skips the download from the transition
database servers completely when it's ignored.

On my VM, the transition database took c. 45mins to replicate (a fair
chunk of that was the script waiting on postgres to create the indexes
after loading all the data).