Increase default elasticsearch local proxy timeout
We're seeing 504 gateway timeout errors in some of our batch jobs.  Some
of these jobs bulk load large amounts of data in a single request, and
therefore expect these requests to take a significant amount of time.
Setting an appropriate timeout should therefore be the responsibility of
the client application which is in a position to know what sort of
request it's making (eg rummager sets a different timeout for bulk and
interactive requests [1]).

This PR therefore increases the default read timeout for these proxy
vhosts to 60s (which is the nginx default value).

[1] https://github.com/alphagov/rummager/blob/master/lib/elasticsearch/index.rb#L55-L64