Add support for RECORDFAIL_TIMEOUT using Puppet
GRUB2 has the concept of RECORDFAIL, which sets a series of behaviours to
apply on the boot after a failed boot. When a machine boots in to this
RECORDFAIL behaviour, it will always stick at the GRUB2 boot menu. Setting
the TIMEOUT value stops this from stalling boot of a headless system.

The GRUB2_TIMEOUT value applies on every boot; the GRUB2_RECORDFAIL_TIMEOUT
value applies only when the RECORDFAIL behaviour occurs first.