Don't raise CRITICAL alerts for reboot required
Set the threshold for "reboot required by apt" to be one year so that it
only raises a WARNING instead of CRITICAL, unless we've put it off for an
incredibly long period of time. It's not possible to give a value that
indicates never and it doesn't seem worth modifying the script.

These alerts are never immediately actionable because they occur for
machines that are single points of failure and excluded from unattended
reboots, which means that they distract 2nd line from real alerts on the
dashboard.

This is band-aid for [#96769468] which describes how we might want to
consolidate or separate these alerts.