Fix slow DNS lookups on trusty machines
Trusty machines were taking a long time to resolve dns requests when
using curl to make a http request. This is because the dns lookup
process makes an ipv4 and ipv6 request, and waits for the ipv6 response
(which never happens) and times out, only then returning the ip address
returned by the ipv4 lookup

From the resolv.conf manpage -

"The resolver uses the same socket for the A and AAAA requests.  Some
hardware mistakenly sends back only one reply.  When that happens the
client system will sit and wait for the second reply.  Turning this
option on changes this behavior so that if two requests from the same
port are not handled correctly it will close the socket and open a new
one before sending the second request."